backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
chromadb==0.4.22
langchain==0.1.5
langchain-community==0.0.17
dashscope==1.14.1
apscheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.41.1
lxml==5.1.0
python-dateutil==2.8.2
